Folder restructure and Player now damages enemies
Added a folder to the scripts folder for scripts that the player uses and one for the Cube Grid as well.

For the player's cannons, they now deal damage instead of destroy the enemy on hitting them, just need to have the parts of the enemy ship to send damage to the overall ai.